Port scalar GDN Mega kernels (#427)
## Human Note

## Agent note

Port NVIDIA Frost's BT64 scalar-GDN prefill, checkpoint-recompute, and bprop kernels onto the
landed Mega runtime without routing scalar gates through KDA's bounded reciprocal rebase. This
layer includes the private Torch launchers, source attribution, terminal-safe packed TMA maps,
K-first checkpoint descriptors over public [V, K] storage, end-to-end int64 descriptor strides,
and the shared autograd-context CUDA device guard used by GDN and KDA.

The private tests are intentionally part of the kernel-port layer. They cover every BT64 boundary,
packed second sequences and empty sequences, poisoned terminal capacity, checkpoint orientation,
FP16/BF16, extreme and model-style gates, grouped-head gradient reduction, outer strides,
misalignment rejection, forced int64, and active Q/K addresses beyond signed int32. CUTracer
random-delay stress at 5, 20, and 100 microseconds remained numerically correct and bitwise
repeatable for prefill, recompute, bprop, and their prologues.

## Test Plan

```bash
gpu-run auto -- uv run --extra mega --with pytest pytest \
  test/gdn/mega/test_gdn_mega_forward.py \
  test/gdn/mega/test_gdn_mega_backward.py \
  test/gdn/mega/test_gdn_mega_layouts.py \
  test/kda/mega/test_kda_mega_training.py::test_mega_forced_int64_forward_backward_matches_int32 -q

# CUTracer oracle was run for forward and backward at delay-ns 5000, 20000, and 100000.
```

+++ attn_gym/linear/_delta_rule/mega/kernels/common/thd.py
@@ -31,7 +31,7 @@ def emit_seq_load_descs(
     cu_seqlens,
     base_ptr,
     n_batch: cutlass.Int32,
-    row_stride: cutlass.Int32,
+    row_stride: cutlass.Int64,
     seq_ord: cutlass.Constexpr[int],
 ) -> None:
     """Build sequence-relative input maps except at the physical token endpoint.
@@ -82,7 +82,7 @@ def emit_seq_descs(
     cu_seqlens,
     base_ptr,
     n_batch: cutlass.Int32,
-    row_stride: cutlass.Int32,
+    row_stride: cutlass.Int64,
     seq_ord: cutlass.Constexpr[int],
 ) -> None:
     """Per-BATCH TMA-descriptor array for a VARLEN (THD) tensor whose base
@@ -126,7 +126,7 @@ def emit_checkpoint_seq_descs(
     cu_seqlens,
     base_ptr,
     n_batch: cutlass.Int32,
-    row_stride: cutlass.Int32,
+    row_stride: cutlass.Int64,
     every_n: cutlass.Int32,
     seq_ord: cutlass.Constexpr[int],
 ) -> None:
